try for best slipt for code structuration
